feat: Add hash package for generating SHA1 hashes
The code changes introduce a new package called `hash` which provides a function `New` for generating SHA1 hashes from a byte slice. This package uses the `crypto/sha1` and `encoding/hex` packages from the Go standard library.

This commit message follows the established convention of using a prefix to indicate the type of change. In this case, "feat" is used to indicate a new feature or functionality being added.

Please remove any additional information and wrap the commit message in a single markdown code block.